apply review fixes
