+++ packages/fuels-code-gen/src/program_bindings/abigen/bindings/script.rs
@@ -52,11 +52,11 @@ pub(crate) fn script_bindings(
                 }
             }
 
-            pub fn with_account<U: ::fuels::accounts::Account>(&self, mut account: U) -> ::fuels::types::errors::Result<#name<U>> {
+            pub fn with_account<U: ::fuels::accounts::Account>(self, mut account: U) -> ::fuels::types::errors::Result<#name<U>> {
                 let provider = ::fuels::accounts::ViewOnlyAccount::try_provider(&self.account)?;
                 account.set_provider(provider.clone());
 
-               ::core::result::Result::Ok(#name { account, binary: self.binary.clone(), log_decoder: self.log_decoder.clone()})
+               ::core::result::Result::Ok(#name { account, binary: self.binary, log_decoder: self.log_decoder})
             }
 
             pub fn with_configurables(mut self, configurables: impl Into<::fuels::programs::Configurables>)
